+++ Assets/Scripts/GunManager.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 1955175cd1e8b46d497323ab442c384f
-timeCreated: 1496547274
+guid: ed985227d6a3b4c8493c7caf639e9316
+timeCreated: 1496649272
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
