+++ Assets/Scripts/RayControl.cs
@@ -5,7 +5,6 @@ using UnityEngine;
 public class RayControl : MonoBehaviour {
 	public GameObject sparkle;
 	public Renderer sparkle2;
-    public GameControl gameControl;
 	GameObject bullet;
 
 
@@ -16,14 +15,9 @@ public class RayControl : MonoBehaviour {
 	
 	// Update is called once per frame
 	void Update () {
-		if(gameControl.shot == true){
-			GenerateBullet();
-		}else if(gameControl.shot == false){
-			DestroyBullet(); 
-		}
 	}
 
-	void GenerateBullet(){
+	public void GenerateBullet(){
 		sparkle2.enabled = true;
 		Ray ray = Camera.main.ScreenPointToRay(Input.mousePosition);
 		Debug.DrawRay(ray.origin, ray.direction, Color.red, 3.0f);
@@ -34,7 +28,7 @@ public class RayControl : MonoBehaviour {
 		
 	}
   
-  void DestroyBullet(){
+  public void DestroyBullet(){
 		sparkle2.enabled = false;
 		Destroy(bullet);       
   }
